+++ CHANGELOG.rst
@@ -5,6 +5,16 @@ Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to `Semantic Versioning <http://semver.org/>`_.
 
+0.2.1-dev
+*********
+
+Incompatible changes
+--------------------
+
+* Data::key() now returns std::string instead of DataConstRef.
+* Data::value() now return Data rather than DataConstRef
+
+
 0.2.0
 *****